requirements.txt
Flask==2.0.1
gunicorn==20.1.0
requests==2.26.0
PyJWT==2.4.0
pydub==0.25.1
Werkzeug<2.1.0
numpy==1.23.5
moviepy==1.0.3
Pillow==10.0.0
pysrt==1.1.2
openai-whisper==20230314
deep-translator==1.11.4